Roster保存了版本号，但是现在有个地方会报异常，虽然没有蹦

+++ openIM/src/main/java/com/open/im/db/OpenIMDao.java
@@ -53,8 +53,10 @@ public class OpenIMDao {
      * @param list 成组的VCardBean
      */
     public void saveAllVCard(Collection<VCardBean> list) {
-        deleteAllVCard();
-        DataSupport.saveAll(list);
+        if (list != null){
+            deleteAllVCard();
+            DataSupport.saveAll(list);
+        }
     }
 
     /**
